feat: Implement MVI base for Home screen and core components

+++ feature/home/src/main/java/com/uzunguc/financetracker/feature/home/ui/Operation.kt
@@ -1,11 +1,12 @@
 package com.uzunguc.financetracker.feature.home.ui
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class Operation(
     val id: Int,
     val name: String,
-    val sum: Long,
+    val sum: BigDecimal,
     val type: TypeOperation,
     val date: LocalDateTime
 )
